fix simulazione 2 run.sh

+++ simulazioni/2/run.sh
@@ -1,5 +1,10 @@
 #!/bin/bash
 
+if [ $# -ne 2 ]; then
+  echo "Errore: devono essere passati esattamente due argomenti."
+  exit 3
+fi
+
 if [[ -e $1 ]]; then
     rm "$1"
 fi
